Final version as demonstrated in the practical
This is the Finial version that was shown in the demonstration. It's got
bugs. Lacks input verification, so be careful with input.
The application needs to be shutdown through ctrl+alt+del, because the
WebServer thread on;y shuts down if the application is exited through
the main window.
Have fun.